Rename content files
