+++ dbaccess/specs.py
